requirements.txt
streamlit
opencv-python-headless
mediapipe
numpy
pandas
scikit-fuzzy
pillow
